+++ examples/basic/src/model-provider/ollama/ollama-completion-generate-text-example.ts
@@ -1,23 +1,21 @@
 import dotenv from "dotenv";
-import { ollama, streamText } from "modelfusion";
+import { generateText, ollama } from "modelfusion";
 
 dotenv.config();
 
 async function main() {
-  const textStream = await streamText(
+  const text = await generateText(
     ollama
-      .TextGenerator({
+      .CompletionTextGenerator({
         model: "mistral:text",
-        maxGenerationTokens: 500,
+        maxGenerationTokens: 120,
       })
       .withTextPrompt(),
 
     "Write a short story about a robot learning to love:\n\n"
   );
 
-  for await (const textPart of textStream) {
-    process.stdout.write(textPart);
-  }
+  console.log(text);
 }
 
 main().catch(console.error);