+++ src/model-provider/ollama/index.ts
@@ -4,4 +4,4 @@ export * from "./OllamaChatPromptTemplate.js";
 export { OllamaErrorData } from "./OllamaError.js";
 export * as ollama from "./OllamaFacade.js";
 export * from "./OllamaTextEmbeddingModel.js";
-export * from "./OllamaTextGenerationModel.js";
+export * from "./OllamaCompletionModel.js";